backend/requirements.txt
# Django Core
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Image & Video Processing
Pillow==10.1.0
opencv-python==4.8.1.78

# Deep Learning
torch==2.1.1
torchvision==0.16.1
ultralytics==8.0.196

# Data Processing
numpy==1.24.3
pandas==2.1.3

# Async Processing (Optional - choose one)
celery==5.3.4
redis==5.0.1

# WebSocket Support
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# Production Server
gunicorn==21.2.0

# Database (PostgreSQL - Optional)
psycopg2-binary==2.9.9

# Utilities
python-dotenv==1.0.0